addon/requirements.txt
typing>=3.7.4.3
fake-bpy-module-2.90>=20200907
lxml~=4.5.2